==> 3. Augment data with Zefix.ipynb <==
import traceback

import nltk
from nltk.corpus import stopwords
import re

import pandas as pd

from helpers.zefix_scraper import scrape_company

from helpers.language import stringify_list, sanitize_pipeline

def string_pipeline(s):
    return stringify_list(sanitize_pipeline(s))

interests = pd.read_json('data/all_interests.json')
interests['interest_name_clean'] = interests.interest_name.apply(string_pipeline)
interests = interests.sort_values(by="index")
interests.head(5)

scraped_companies = []
failed = []
for _, row in interests.iterrows():
    try:
        data = scrape_company(row.interest_name_clean) 
        if data is None:
            continue
        data["source"] = {
            "first_name": row.first_name,
            "last_name": row.last_name,
            "interest_name": row.interest_name,
            "data_source": row.data_source,
            "index": row["index"]
        }
        print(row["index"])    
        scraped_companies.append(data)
    except Exception as e:
        print("Error: " + str(e))
        traceback.print_exc()
        failed.append(row["index"])


open("scraped_companies_2.json", "w").write(json.dumps(scraped_companies2))

len(scraped_companies2)
