==> 4. Lookup politicians and their interests in the Panama Papers.ipynb <==
# Lookup politicians and their interests in the Panama Papers

Using the data we obtained in the previous notebooks, we will now look up those politicians and their interests in the panama papers.

%matplotlib inline
import json
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
# New imports for using Neo4j
import networkx as nx
#from py2neo import Graph
from helpers.vis import draw
%set_env NEO4J_URL http://neo4j:neo4j2@localhost:7474
#%load_ext cypher

graph = Graph('http://neo4j:neo4j2@localhost:7474')

# function to search columns of a pandas dataframe
def search(df, col, words):
    cond = [df[col].str.contains('(?<![\w\d])'+w+'(?![\w\d])', case=False, na=False) for w in words]
    return df[np.logical_and.reduce(cond)]

# split a string in words with no ponctuation
import re
def split(str):
    return re.split('\W+', str)

# load datasets
from helpers.swiss_codes import filter_swiss
# 1. officers
officers = pd.read_csv('data/Officers.csv', header=0, low_memory=False)
swiss_officers = filter_swiss(officers, process_address=False)
del officers
# 2. entities
entities = pd.read_csv('data/Entities.csv', header=0, low_memory=False)
swiss_entities = filter_swiss(entities)
del entities
# 3. intermediaries
inter = pd.read_csv('data/Intermediaries.csv', header=0, low_memory=False)
swiss_inter = filter_swiss(inter)
del inter

## All interests

all_interests = pd.read_json('data/all_interests.json')
all_interests

Looking for names in Officers:

rs = []
for _,f,l in all_interests[['first_name','last_name']].drop_duplicates().itertuples():
    ws = f + ' ' + l
    r = search(swiss_officers, 'name', split(ws))
    if r.size > 0 or l == 'Zacharias':
        rs.append(ws)

rs

And the companies from Entities and Intermedietaries:

try:
    inters = pd.read_csv('data/all_interestsd_entities_name.csv', index_col=0)
except:
    ents = []
    for _, interest in all_interests.drop_duplicates(subset=["interest_name"]).iterrows():
        ws = split(interest["interest_name"])
        r = search(swiss_entities, 'name', ws)
        if r.size > 0:
            for idx,name in r['name'].iteritems():
                ents.append([interest["interest_name"].strip(),idx,name,interest["data_source"],interest["first_name"],interest["last_name"]])
    inters = pd.DataFrame(ents, columns=['interest_name','idx','entity_name', "data_source", "first_name", "last_name"])
    inters.to_csv('data/all_interests_entities_name.csv')

inters

scraped_interests = json.loads(open("scraped_companies_1.json").read())
scraped_interests2 = json.loads(open("data/scraped_companies.json").read())

second_order_people = []

for company in scraped_interests+scraped_interests2:
    if not "persons" in company:
        continue
    for person in company["persons"]:
        second_order_people.append({
                "first_name": person["first_name"],
                "last_name": person["last_name"],
                "source": company["source"]
            })
        
second_order_people = pd.DataFrame(second_order_people)

second_order_results = []

for _, first, last, source in second_order_people.drop_duplicates(subset=["first_name", "last_name"]).itertuples():
    full_name = first + " " + last
    results = search(swiss_officers, 'name', split(full_name))
    if results.size > 0:
        for idx,name in results['name'].iteritems():
            second_order_results.append([full_name,idx,name, source["first_name"], source["last_name"], source["data_source"], source["interest_name"]])
    #second_order_results = pd.DataFrame(second_order_results, columns=['interest_name','idx','entity_name', 'source'])
    


pd.DataFrame(second_order_results)

These matches have to be checked more in details as they seem to mean something else...

## Parlement data

parlement = pd.read_json('data/parliament_members_interests.json')

ps = []
for _,f,l in parlement[['FirstName','LastName']].drop_duplicates().itertuples():
    ws = f + ' ' + l
    r = search(swiss_officers, 'name', split(ws))
    if r.size > 0:
        ps.append(ws)

ps

Only some matches and there're a lot of possibilities with these names...

### Querying the Neo4j with Poggia

%%cypher 
MATCH (o:Officer) WHERE toLower(o.name) CONTAINS "poggia"
RETURN o.name as name, o.countries as countries LIMIT 20

# Plotting Poggia's graph
results = %cypher \
    MATCH (o:Officer) WHERE toLower(o.name) CONTAINS "poggia" \
    match p=(o)-[r*1..100]-() \
    RETURN p limit 100

G = results.get_graph()
draw(G)

# Plotting energo graph
results2 = %cypher \
    MATCH (o1:Entity) WHERE toLower(o1.name) CONTAINS "energo" \
    match p=(o1)-[r*1..100]-() \
    RETURN p limit 100

G2 = results2.get_graph()
draw(G2)

We can see results in Morges and Porrentruy (+ Russia, Spain, Panama) all centralized around GENINT SA  
http://ge.ch/hrcintapp/externalCompanyReport.action?companyOfrcId13=CH-660-0806985-5&ofrcLanguage=2

# Plotting a graph concerning the IMD school in Lausanne ;)
results3 = %cypher \
    MATCH (o1:Address) WHERE toLower(o1.address) CONTAINS "schmidheiny" \
    match p=(o1)<-[r*1..100]-() \
    RETURN p limit 20

G3 = results3.get_graph()
draw(G3)
